+++ src/main/java/com/trego/dto/PreOrderDTO.java
@@ -2,7 +2,6 @@ package com.trego.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -12,8 +11,8 @@ public class PreOrderDTO {
     private long orderId;
     private long addressId;
     private String mobileNo;
-    private BigDecimal totalCartValue;
-    private BigDecimal amountToPay;
+    private Double totalCartValue;
+    private Double amountToPay;
     private double discount;
     private List<CartDTO> carts;
