+++ src/main/java/com/trego/dto/ProductDTO.java
@@ -1,15 +1,14 @@
 package com.trego.dto;
 
 import lombok.Data;
-import java.math.BigDecimal;
 
 @Data
 public class ProductDTO {
     private Long id;
     private String name;
-    private BigDecimal price;
-    private BigDecimal tax;
-    private BigDecimal totalPrice;
+    private Double price;
+    private Double tax;
+    private Double totalPrice;
     private String description;
     private String image;
     private Integer stock;
