+++ src/main/java/com/trego/dto/SubstituteDetailDTO.java
@@ -1,7 +1,6 @@
 package com.trego.dto;
 
 import lombok.Data;
-import java.math.BigDecimal;
 
 @Data
 public class SubstituteDetailDTO {
@@ -11,7 +10,7 @@ public class SubstituteDetailDTO {
     private String manufacturer;
     private String vendorName;
     private String vendorLogo;
-    private BigDecimal mrp;
-    private BigDecimal bestPrice;
-    private BigDecimal discount;
+    private Double mrp;
+    private Double bestPrice;
+    private Double discount;
 }
