+++ src/main/java/com/trego/dto/response/CartResponseDTO.java
@@ -4,24 +4,23 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.trego.dto.MedicineDTO;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class CartResponseDTO {
     private Integer vendorId;
-    private BigDecimal totalCartValue;
-    private BigDecimal amountToPay;
-    private BigDecimal discount;
+    private Double totalCartValue;
+    private Double amountToPay;
+    private Double discount;
     private  long orderId;
     private String name;
     private String licence;
     private String gstNumber;
    // private String address;
     private String logo;
-    private BigDecimal lat;
-    private BigDecimal lng;
+    private Double lat;
+    private Double lng;
     private Integer deliveryTime;
     private String reviews;
     private String rating;
