+++ src/main/java/com/trego/dto/response/OrderItemDTO.java
@@ -1,19 +1,17 @@
 package com.trego.dto.response;
 
-
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 
 @Data
 public class OrderItemDTO {
-   private  BigDecimal totalAmount;
+   private Double totalAmount;
    private long itemId;
-   private  int qty;
-   private BigDecimal mrp;
-   private BigDecimal price;
+   private int qty;
+   private Double mrp;
+   private Double price;
    private Map<String, Object> medicine = new HashMap<>();
 
 }
